Revert "Test against Pinot 1.1.0"
The new TestPinotPreviousConnectorTest is too flaky.

This reverts commit d5502607d681dad075a16abf36694074042e6d76.
