+++ plugin/trino-pinot/src/test/java/io/trino/plugin/pinot/TestingPinotCluster.java
@@ -67,7 +67,6 @@ public class TestingPinotCluster
         implements Closeable
 {
     public static final String PINOT_LATEST_IMAGE_NAME = "apachepinot/pinot:1.2.0";
-    public static final String PINOT_PREVIOUS_IMAGE_NAME = "apachepinot/pinot:1.1.0";
     private static final String ZOOKEEPER_INTERNAL_HOST = "zookeeper";
     private static final JsonCodec<List<String>> LIST_JSON_CODEC = listJsonCodec(String.class);
     private static final JsonCodec<PinotSuccessResponse> PINOT_SUCCESS_RESPONSE_JSON_CODEC = jsonCodec(PinotSuccessResponse.class);
